+++ Bloch/docs/release-notes.md
@@ -4,8 +4,9 @@
 
 ## Upcoming release
 
-* Removed the Attention Span, Calm, Emotional Damage and Over-stimulation cards.
+* Removed the Attention Span, Calm, Emotional Damage, Over-stimulation and Psionic Shield cards.
 * Removed the Unlocked Potential artifact.
+* Added the Unlocked Potential card.
 * Reworked the Outburst card.
 * Buffed the Focus card.
 * Technically buffed the Reality Bending card by making it Scry first.